Created method to execute report generation
Completed successful manual testing of generating report by processing files saved on system and generating new output file using transformation rule saved in database, The task can be triggered manually else executed by amqp at the given schedule time

+++ src/main/java/com/service/report/generator/utility/TextWizard.java
@@ -34,12 +34,12 @@ public class TextWizard implements TextHelper{
     }
 
 
-    public float evaluateExpression(String expression) {
+    public double evaluateExpression(String expression) {
         try (Context context = Context.newBuilder()
-                .option("engine.WarnInterpreterOnly", "false") // Disable warning
+                .option("engine.WarnInterpreterOnly", "false")                  // Disable warning
                 .build()) {
             Value result = context.eval("js", expression);
-            return result.asFloat();
+            return result.asDouble();
         }
     }
